Remove old tests and fixtures
